cli: add a hidden flag to turn on klog traces, which logs all k8s api calls
