+++ internal/cli/klog.go
@@ -1,0 +1,36 @@
+package cli
+
+import (
+	"flag"
+	"fmt"
+	"log"
+	"os"
+
+	"k8s.io/klog"
+)
+
+var klogLevel = 0
+
+// HACK(nick): The Kubernetes libs we use sometimes use klog to log things to
+// os.Stderr. There are no API hooks to configure this. And printing to Stderr
+// scrambles the HUD tty.
+//
+// Fortunately, klog does initialize itself from flags, so we can backdoor
+// configure it by setting our own flags. Don't do this at home!
+func initKlog() {
+	var tmpFlagSet = flag.NewFlagSet(os.Args[0], flag.ExitOnError)
+	klog.InitFlags(tmpFlagSet)
+
+	flags := []string{
+		"--stderrthreshold", "FATAL",
+	}
+
+	if klogLevel > 0 {
+		flags = append(flags, "-v", fmt.Sprintf("%d", klogLevel))
+	}
+
+	err := tmpFlagSet.Parse(flags)
+	if err != nil {
+		log.Fatal(err)
+	}
+}
